+++ pkg/parca/parca.go
@@ -76,10 +76,12 @@ const (
 )
 
 type Flags struct {
-	ConfigPath         string   `default:"parca.yaml" help:"Path to config file."`
-	Mode               string   `default:"all" enum:"all,scraper-only" help:"Scraper only runs a scraper that sends to a remote gRPC endpoint. All runs all components."`
-	LogLevel           string   `default:"info" enum:"error,warn,info,debug" help:"log level."`
-	Port               string   `default:":7070" help:"Port string for server"`
+	ConfigPath  string `default:"parca.yaml" help:"Path to config file."`
+	Mode        string `default:"all" enum:"all,scraper-only" help:"Scraper only runs a scraper that sends to a remote gRPC endpoint. All runs all components."`
+	LogLevel    string `default:"info" enum:"error,warn,info,debug" help:"log level."`
+	HTTPAddress string `default:":7070" help:"Address to bind HTTP server to."`
+	Port        string `default:"" help:"(DEPRECATED) Use http-address instead."`
+
 	CORSAllowedOrigins []string `help:"Allowed CORS origins."`
 	OTLPAddress        string   `help:"OpenTelemetry collector address to send traces to."`
 	Version            bool     `help:"Show application version."`
@@ -131,6 +133,11 @@ func Run(ctx context.Context, logger log.Logger, reg *prometheus.Registry, flags
 			return err
 		}
 		defer closer()
+	}
+
+	if flags.Port != "" {
+		level.Warn(logger).Log("msg", "flag --port is deprecated, use --http-address instead")
+		flags.HTTPAddress = flags.Port
 	}
 
 	cfg, err := config.LoadFile(flags.ConfigPath)
@@ -409,7 +416,7 @@ func Run(ctx context.Context, logger log.Logger, reg *prometheus.Registry, flags
 			return parcaserver.ListenAndServe(
 				ctx,
 				logger,
-				flags.Port,
+				flags.HTTPAddress,
 				flags.CORSAllowedOrigins,
 				flags.PathPrefix,
 				server.RegisterableFunc(func(ctx context.Context, srv *grpc.Server, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) error {
@@ -603,7 +610,7 @@ func runScraper(
 				return parcaserver.ListenAndServe(
 					serveCtx,
 					logger,
-					flags.Port,
+					flags.HTTPAddress,
 					flags.CORSAllowedOrigins,
 					flags.PathPrefix,
 					server.RegisterableFunc(func(ctx context.Context, srv *grpc.Server, mux *runtime.ServeMux, endpoint string, opts []grpc.DialOption) error {