+++ ui/packages/shared/profile/src/useGrpcQuery/index.ts
@@ -11,16 +11,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ui
+import {useQuery, UseQueryResult} from 'react-query';
 
-import "embed"
+interface Props<IRes> {
+  key: string | any[];
+  queryFn: () => Promise<IRes>;
+  options?: {
+    enabled?: boolean | undefined;
+    staleTime?: number | undefined;
+  };
+}
 
-//nolint:typecheck
-//go:embed packages/app/web/build
-var FS embed.FS
+const useGrpcQuery = <IRes>({
+  key,
+  queryFn,
+  options: {enabled = true, staleTime} = {},
+}: Props<IRes>): UseQueryResult<IRes> => {
+  return useQuery<IRes>(
+    key,
+    async () => {
+      return await queryFn();
+    },
+    {
+      enabled,
+      staleTime,
+    }
+  );
+};
 
-// NOTICE: Static HTML export of a Next.js app contains several files prefixed with _,
-// directives for all these patterns need to explicitly added.
-// > If a pattern names a directory, all files in the subtree rooted at that directory are embedded (recursively),
-// > except that files with names beginning with ‘.’ or ‘_’ are excluded.
-// source: https://pkg.go.dev/embed
+export default useGrpcQuery;